Wait on user before rendering the classifier (#457)
* Move API call to user store
Add user.checkCurrent() action to check the currently logged-in user from the user store.

* Add a pending flag to the user store
Add a pending flag to the user store. Hide the classifier while we're checking the current user on first load.

* Display classifier errors
Show the error message on the page when there is a problem rendering the classifier.

* Add user.loadingState
Add a loading state to the user store, consistent with the project store.
Catch any errors during initial sign-in and record them as user.error.

* Use ErrorMessage for project loading errors
Display project loading errors and classifier errors with the ErrorMessage component.