Add onCompleteClassification callback to the classifier (#504)
* Add onComplete callback on classification complete
Add an onComplete callback to the classification store. Call it with the classification and subject on classification complete.

* Add onCompleteClassification prop
Add onCompleteClassification as a callback prop for the Classifier component.

* Add onCompleteClassification to dev classifier

* Update recents on classification complete
Add onCompleteClassification to the classifier wrapper. Add a new recent subject when a classification completes.

* Add tests for the classifier wrapper
Update JSDOM to mock session storage.
Test the classifier with and without a loaded user and project.
Test that the onCompleteClassification callback updates recent subjects.

* lint