Add to collection (#522)
* Add a basic 'Add to collection' modal
Add the basic CollectionsModal component and its container. Connect the container to the collections store.
Sketch out initial expectations for the collections form.

* Wire up select a collection and add subjects
Wire up the select input to search collections. Disable the Add button until a collection is selected. On add, add new subjects to the selected collection.
Sketch out the CreateCollection component and tests.

* Fix add/remove subjects
Use id, not collectionId.

* Wire up CreateCollection
Connect up CreateCollection with the store actions. Disable the form unless you've entered a collection name. Store the form state in CollectionsModalContainer. Add tests for CreateCollection props and CollectionsModalContainer state.

* Close on submit
Add a close() method to the modal container and close after adding subjects to a collection.

* Remove Grommet from the collections modal

* Use forms for collections search and create
Use forms, with submit buttons, to interact with collections.
Restrict collections search by owner/collaborator permissions.
Update tests.

* Remove console logs

* Move modal to classify page

* Wire up the Add To Collection button
Add an addToCollection callback prop to the classifier.
On click of the collections button, call the callback and open the collections modal with the current subject ID.
Add the subject ID to the API calls for creating or add to collections.

* Show the selected collection name
Set the Select value after selecting a collection, so that the collection name is displayed in the text box.

* Move grid display to forms
Remove Grid from the modal. Add grids to the form contents instead.