Disable task area while subject is loading (#791)
* Add an onReady prop to subject viewers
Add an onReady prop and call it when the subject is loaded and ready to classify.

* Add ready state to subject viewer
Add subject ready state to the subject viewer store, and reset it on each new subject.

* Set subject viewer ready state on load
Add an action to set ready state when subject is fully loaded/. Wire it up to the onReady handler for subjects.

* Add subject ready state to tasks
Add subject viewer ready state to tasks. Pass a disabled prop down to task components and task buttons. For now, since disabled isn't implemented on those components, hide the task area until the subject is ready.

* Disable task nav buttons
Add a disabled prop to the task nav button. Disable them while a subject is loading.

* Add disabled to TaskInputField
Add a disabled prop to task input fields, and disable the input element when disabled.
Remove hover styles when disabled.

* Add disabled prop to question tasks
Allow multiple and single answer questions to be disabled.

* Call onReady on image load
Pass onReady down to the single image viewer as the image onLoad handler.
Test that the load event calls onReady.

* LCV: call onReady after parsing subject data
Call onReady in the light curve viewer after the subject data has been parsed by D3.

* Rename resetSubject

* Add tests
Test that task components are disabled until the subject viewer is ready for use.