Record subject image dimensions on load (#804)
* Record subject image dimensions on load
Get image dimensions from the HTML preloader. Create a new, fake event and fake target on SVG image load, since the original is read-only. Pass the fake event, with the HTML image dimensions, to the subject viewer store with the onReady event handler.

* Enable lifecycle method testing
Allow enzyme shallow rendering to run React lifecycle methods.

* Test the single image preloader
Mock window.Image() and expect onReady to be called with the original image dimensions on image load.

* Fix data subjects
Data subjects don't fire a load event, so height and width are undefined. Ignore naturalHeight and naturalWidth for these subjects, since those numbers are meaningless.

* Add subject dimensions to subject viewer store
Copy the subjectDimensions model from the classificationMetadata model: dimensions is an array of size objects for different frames of the subject.

* Add subject dimensions to the classification

* Reset subject dimensions on subject change

* Add subject viewer to classification store tests

* mock subject dimensions in classification tests
Expect subject dimensions to be copied from the subject viewer to the classification store when submitting a classification.

* Tidy up onSubjectReady

* Record client dimensions for single image subjects
Create a ref to the svg element for single image subjects. Add the client height and width of the svg element to the recorded subject dimensions.

* Record client dimensions for light curve viewer
Get a reference to the light curve viewer's container DOM node, then get client height and width from that node.

* Use a forward ref for the light curve viewer
Pass a ref to the LCV container node as a forward ref to the LightCurveViewer component.

* Use toJS() instead of toJSON()

* Default natural image dimensions to 0