Fix Done And Talk when feedback isn't present (#901)
* Create an openInTalk action for subjects
Add an openInTalk action to subjects and invoke it when Done & Talk is pressed. Remove the code from the DoneAndTalkButtonContainer that updated window.location to open a Talk page. Move tests from DoneAndTalkButtonContainer to the Subject model.

* Always go to Talk from Done & Talk
Don't defer subject.openInTalk() but invoke it immediately that the button is pressed.

* Defer opening Talk until the subject advances
Set a flag when we press Done & Talk, then interrupt subject.advance() and go to Talk if that flag is set. Move code for opening a new Talk page into the subject store and update tests.

* Ignore feedback subjects
Don't go directly to Talk for feedback subjects. Always advance the subject queue, even if opening a Talk page.