Fix shadowing in the SubjectStore tests (#1045)
* Fix shadowing in the SubjectStore tests
Make sure every declaration of rootStore is locally scoped to a function.
Add a test to check that advancing the queue reduces its size by one.

* Fix shadowing when refilling the queue

* Fix subject IDs
The first mock subject ID is '10', not '1'. Fix the first test to look for the IDs that were used in the mock queue.

* Remove .only

* Use beforeEach to record initial state
Use a beforeEach function to record initial state and advance the queue by one. Test final state in the it() functions.

* Rename firstSubject to previousSubject