Adjust preview thumbnail size (#1153)
* Adjust preview thumbnail size
Set the maximum subject thumbnail size to 270px width, 200px height.

* Crop subject previews
Align subject previews in the middle of the preview window. Wrap them in a box with max height and max width to prevent them from becoming too tall or too wide.
Add tests for the maxHeight and maxWidth props on the styled box that frames the media.

* Bump the thumbnail size to 700px
Thumbnails should fill the width of the screen on tablets and phones.