Add Your stats to the Classify page (#1128)
* Add the stats widget back to the classify page

* Increment session counts
Increment your session count when you complete a classification.

* get counts for this week, starting Monday
Add calculateWeeklyStats() to the store, which filters the daily counts for this week's counts, starting on Monday (day number 1 in JavaScript.)
Update the chart component to just render the counts as text for the time being.

* Install vx

* Add a basic bar chart
Add a basic vx bar chart with a fixed height and width for the moment.

* Add random counts for testing.

* Add some default props

* Fix tests
Add mock daily counts for one week to the store tests.
Add a mock counts array to the chart container tests.

* Add some theming
Add colours for the histogram bars and the axis lines.

* Reduce y-scale ticks
Only print 3 tick marks for the y-axis (plus zero.)

* Add a formatting function for day names
Add a function that returns the abbreviated label for a given day name.

* Start adding tests
Test that chart renders the daily bars as accessible images.

* Upgrade sinon
Upgrade to 7.4 so that we can use fake timers.

* Fix YourStats store tests
Mock daily counts properly for a week and mock today to be a Tuesday.

* Add tests for the chart container
Test the conversion filter that generates labels for the chart data.

* Test weekly stats
Test that daily stats for this week are generated after a user logs in. Fix some errors in the date calculations that were caught by those tests.

* Fix dates in container tests
JS counts months from 0

* Remove redundant clock
Fake timers are set at the root of the YourStats store tests, so remove them from the counts tests.

* Adjust the chart size
Increase the viewbox height, to avoid cutting off the x-axis labels.

* Improve mock timers
Only override the Date constructor.
Scope mock dates to the tests that use them.

* Lighten tick labels
Make the tick labels light grey but leave other properties more-or-less the same.

* Adjust chart sizing
Set some constants for height and width and add padding around the image.
Adjust the grid column size inside the classify page.

* Start a component story
Add a plain rendering of the classifications chart.

* Remove the random test counts
Don't replace the stats with a random number when the count is 0.

* Localise the chart title

* Add some tests for components
Test the chart for expected child components.

* add required window variable for daily counts

* change the mock data to match the API results

* Rename counts to stats

* Remove curly brackets

* Set the label font

* Add a helper function to container tests
Add a helper to generate mock stats for the chart container tests.
Rename count to stat in the chart container.

* Clamp the chart range to (0, 10)
Limit the chart range so that it doesn't look weird when the daily counts are smaller than 10, or all 0.

* Clarify the weekly stats calculator
Use Monday instead of Sunday. It was confusing me that the code used Sunday to calculate a week starting on Monday.

* Add some interactivity
Make the chart bars focusable.
Add a change of colour on hover/focus.

* Add bar text
Add text labels to each bar, showing the daily count.
Move focus/hover behaviour into a parent group.

* Remove YourStats.dailyCounts
YourStats.dailyCounts has been replaced by YourStats.thisWeek.

* Update colours
Use the global text colours for the axes.
Use the brand colour for the bar fill.

* Make today's count dynamic
Update the chart for today when a classification is submitted.
Update tests.

* Fix the storybook
Add mock total counts to the chart props.

* Set the bar label size
Set the bar label text size to small.
Hide the bar labels from screenreaders, since they are already included in the alt text.

* Test the chart with short bars
Add one very small count to the chart story.