Add recently classified subjects to the Classify page (#1131)
Add a component showing the user's most recently classified subjects, and some related components.

- Add the new component
- Add the collections modal
- Share sub-components with other consuming components
- Update tests
- Add stories

This was a bit of a massive PR, so the full commit message list is below.

## Commits

* Stub out Recent Subjects
Add stub components and a basic story for the recent subjects/classifications component on the Classify page.

* Stretch the classify page grid

* Add subject previews
Add preview thumbnails for each subject.
Update storybook.

* Stub out subject tools
Add a talk link, favourites button and collections button to each subject preview.

* Allow recents to be favourited
Add a favorite flag to the Recent model.
Update components to pass this flag when they add recents.
Set the state of the Favourites checkbox from recent.favorite.

* Add `Recent.toggleFavorite`
Add a `toggleFavorite` action to Recent and wire it up to the Favourites button. Update the Recent Subjects story to make the favourites buttons interactive.

* Add the collections modal
Add the collections modal to the subject preview and wire it up to the collections button.
Update the collections modal with sensible defaults so that it doesn't break the storybook.

* Fix classifier tests
Add favorite flag to the mock subject.

* Add Talk links
Add the project slug to component props.
Link 'Discuss on Talk' to the subject's Talk URL.
Link the subject preview media to the subject's Talk URL.
Update TalkLink tests.

* Add subject preview tests
Test that the linked media is displayed and linked.
Test that the subject Talk, Favourite and Collections buttons exist.

* Adjust maximum subject height

* Wire up the Recent model to the API
Add/remove subjects from favourites when we toggle their favourite state.

* Add a project slug to the Recent Subjects story

* Add stories to test layout
Add stories for projects with portrait images or video subjects.

* Update preview sizing
Stretch the subject preview grid.
Wrap previews in a Box to prevent them overflowing their container.
Move SubjectPreview height and width into props.

* Add a small screen layout
Change the classify page columns depending on viewport size.
Add a size prop which determines how many subjects to show.
Make the recents subjects grid respond to the number of subjects.

* Remove redundant classNames

* Fix favourites checked state

* Check for a logged-in user
Disable collections and favourites if there's noone logged in. Update tests.

* Consistent spelling of favourite

* Rename subject vars for clarity
Rename `checked` to `isFavourite`.
Rename `recent` to `subject`.

* Appease the linter!

* Map recents to subjects
Convert each recent to a mock subject before passing it into SubjectPreview. This should allow the preview component to be used with any subject model, rather than just recents, from Panoptes.

* Add a story for one recent classification

* Update thumbnail sizes
Resize thumbnails to a 270px by 200px box.

* Fix Talk links
Use `subject.id` to generate links. Update tests.

* fix preview alt text

* Test the favourites button state
Test that the favourites checkbox is checked and unchecked properly when clicked.

* Wrap thumbnails in a box
Wrap thumbnails in a box with specified height and width, in order to crop them. Crop by setting `overflow: hidden`.

* Add stories for SubjectPreview
Add stories for landscape images, portrait images and video subjects.
Fix spelling of `toggleFavourite` in the storybook.

* Add a storybook knob for logged-in users
Add a boolean switch to toggle between logged-in or logged-out users in the stories.

* Move SubjectPreview to `src/shared`
Share SubjectPreview since it will be needed by Recents and Collections.

* Move CollectionsModal to `src/shared`

* Update the SubjectPreview story

* Use CSS height and width
Use CSS height and width to specify the dimensions of subject thumbnails. This should improve the layout on smaller screens.

* Add ResponsiveContent
Change the subject preview height based on viewport size.

* Always show 3 columns in wide viewports
Pass recents size down to the RecentSubjects component and always show that many columns in the grid. Limit column content to 100% width for responsive sizing.

* Fix US spelling of favourite
Add separate language files for US and UK spellings.

* Fix Favourites button
Add `subject.toggleFavourite` to allow favouriting to work.

* Use recents size to calculate preview height
Give up on fresnel, which breaks grid layouts in the storybook by inserting extra, empty columns into the DOM. Instead, we use the size prop to get the height.

* Add placeholder images
Add presentational placeholder images when the recents array is smaller than size.

* Allow the storybook to serve files from `./`
I'm not wildly happy about this, but NextJS doesn't allow you to change the static files directory or URL. https://nextjs.org/docs#static-file-serving-eg-images

* Crop placeholders
Wrap placholders in a Box with the same height and width as subjects, and `overflow: hidden`.

* Centre placeholders in the enclosing Box

* Add a story for phones
Add a story with only one subject for phone screens.

* Add a dark theme toggle to stories
Add a dark theme toggle to Recent Subjects and Subject Preview in the storybook.

* Add viewports to all Recent Subjects stories

* Add a placeholder to subject previews
Add a placeholder prop to the subject preview component, and use the Zooniverse logo image as the placeholder for recent subjects.