Add a multi-stage Docker build for production apps (#1109)
* Add a docker production build stage
Split the docker build into bootstrap, which installs the monorepo, and production-apps, which builds the NextJS apps.
Remove the NextJS builds from the bootstrap script.

* Update monorepo docker-compose
Rename the dev service to yarn. Remove local volumes and run tests on the installed image instead. Run test:ci instead of test.
Add a production image to build the project and content apps.
Add named targets to the builds.

* Add named targets to docker-compose
Build the dev apps from the bootstrap target, or the dev image if it already exists.

* Ignore Dockerfile in builds
Changes in the Dockerfile shouldn't change built images.

* Update READMEs

* Update the Jenkins build
Build zooniverse/front-end-monorepo with the bootstrapped monorepo, then build and start the NextJS apps in their respective containers on kubernetes.