Package up the favourites button (#1200)
* Package up the favourites button
Make the favourites button self-contained by moving the button's checked state into a useState hook. Set the initial checked state from subject.favorite. Update tests by checking that the button icon changes when it is checked and unchecked.

* Fix icon tests
The filled prop is boolean, not a string.

* Add documentation
Add a brief README and a storybook story.

* Fix story name