Add a gradient background to workflow links (#1204)
* Add a gradient background to workflow links
Pass the workflow completeness to workflow buttons, and render a background gradient based on that value.

* Add an expanded label for screenreaders
Announce the workflow name and completeness to screenreaders.

* Add hover styles
Mute the gradient sightly on hover or focus. Keep the text colour as black, always.

* Add a white divider to the progress bar
Convert gradients to arrays, for convenience, and add a white 1px divider. Update gradients to match the latest design.

* Format the workflow link labels
Add the completeness percentage to the workflow labels and range left. Remove the ARIA label, which is just duplicating the button label now.
Note that % has to be escaped as %% in the locale files.

* Fix percentages
Set the radix for the percentage calculation, and use the calculated percentage in the gradient style.

* Run the linter

* Add workflow.completeness to props

* Add workflow selector stories
Mock the NextJS router to render test workflow links. Add stories for loading, errored and loaded workflows.

* Add router.query to prop definitions

* Set workflow button text colour
Always use black text, even in the dark theme.

* Align text in the theme
Range the workflow button label left.