Package up the collections button (#1202)
* Package up the collections button
Make the collections button self-contained by adding a collections modal to it. Open the modal when the button is clicked (the modal is responsible for closing itself.) Add tests to check that the modal is opened when the button is enabled and clicked.

* Add documentation
Add a brief README and a storybook story.