Move favourites button to shared react components (#1218)
* Move favourites button to shared react components
Move FavouritesButton and MetaToolsButton to the react-components library. Update the classifier and project app to use the shared buttons.

* Remove the filled attribute
Pass an SVG fill colour down to the HeartIcon component, rather than the boolean filled attribute.
Update tests to use a mock theme and test the fill attribute of the heart icon.