Preload images for the single image viewer (#1236)
* Add stories for SingleImageViewer

* Remove onLoad, onError from SIngleImageViewer

* Preload subject images
Preload subject images with a HTML <img> element. Store the loaded image in component state and render it as an SVG image once loaded.
Update tests.

* Add viewbox to the SVG image
Pass the image height and width to the SVG image and use them to construct the initial viewbox. This should render images full size in Edge.