Make project announcement dismissible (#1235)
Makes the ProjectAnnouncement component dismissible on clicking the close icon.

It works by storing a hash (using @sindresorhus/string-hash) of the announcement 
string as a session cookie, and comparing that hash against one of the current
project announcement. This is done so that a volunteer can't dismiss _all_
announcements: if the string changes, they'll see it again (there's no 
`updated_at` string or similar on the project resource for 
announcement strings to check for changes).

Using a cookie allows us to SSR the page with/without the announcement.

This also styles the banner for wide screens to ensure the text is centred 
despite having a right-aligned close icon. On narrow screens, it's left-aligned.

## Future thoughts

- The logic for this could be moved out of a global store and into the 
ProjectAnnouncementContainer. It's not really global state.
- We could implement a cookie store to handle cookies and abstract out some 
logic around cookie getting/setting.