Add separate layout for ProjectHomePage
This creates a `100vh`-style hero layout for the project home page.

It does this by moving the upper chrome components - site navigation, project navigation and the project announcement - out of `_app.js`, and
moving them into two places. The first is a `StandardLayout` component,
which renders them in the manner we know and love. The second is into
the `ProjectHomePage` component, so they can be composed into a `100vh`
container that fills the viewport.

It also updates the tests and stories to deal with this.