Create an API route for Team page
Creates a Next.js API route to build the data for the team page.
This means we don't have to expose the Contentful access token to the
world on the client side, we can transfer less data to the client,
and we can cache the results.

* Deletes the MST `Team` store
* Rewrites `TeamContainer` to request the data from the API route, and
keep state using hooks
* Add node-cache to cache the result
* Add some tests

**Links**

- https://github.com/zeit/next.js/#api-routes
- https://github.com/node-cache/node-cache

**Future Improvements**

- An API route to manually flush the cache, just in case it gets stuck