Extract annotations out of classification model (#1319)
* Extract annotations out of classification model
Move annotations to an AnnotationsStore model, so that they can be also be used by drawing marks.
Make Classification an AnnotationsStore.

* Add a annotation(task) view
Add a view which returns the annotation for a given task, or the default task annotation if no current annotation exists.

* Add a helper to get annotations by task
Add a helper view to the classifications store, which returns the annotation for a given task. Update the task model to use that view.