Drawing: check for out-of-bounds marks (#1357)
* Drawing: check for out-of-bounds marks
Delete marks when they are completely outside the SVG drawing canvas. Based on PFE's isInBounds function.

Test that passed marks are rendered, and that the active mark is rendered correctly.

Pass the current target to dragEnd and use that to calculate a drawing tools bounds.

Add mock SVG and mark bounds to the tests. Test whether the mark is deleted when inside, overlapping or completely outside the SVG element.