Subject image pan and zoom (#1389)
* Image pan and zoom
Add pan and zoom handlers to the single image viewer.
Add a dynamic viewbox to the single image viewer.
Make subject images focusable and add a keyDown handler.
Add up and down key handlers to keyboard zoom.
Wrap each zoom button in a span which listens for pointer down and starts repeated calls to the zoom callback. Clear the timer on pointer up.