Classifier: Grouped subject models (#1557)
Move Subject model and tests into its own folder.
Update tests to mock the root store and client without throwing errors.
Add GroupSubject, which is a subject composed of SingleImageSubjects. Otherwise, it has the same attributes as Subject.
SingleImageSubject is a refinement of Subject which only has one image URL as a location.
Allow SubjectGroup, SingleImageSubject and Subject as valid models to be stored in the subjects store.