Classifier: fix Tasks storybook errors (#1577)
* Classifier: fix Tasks storybook errors
Tasks stories created a new store each time the story function ran (eg. when a knob setting changes), which throws errors in the MobX Provider. The store is redefined to be a single, constant variable and individual stories change it by calling actions on the WorkflowSteps model.
Some undefined reference errors in individual stores are also caught and fixed here.
Simplify workflowSteps.setSteps.