Classifier: Validate locations before creating subjects (#1575)
Add validateSubjectLocations as a refinement to subject.locations.
Change the validators to return false, rather than throwing, for invalid MIME types and file extensions.
Expect Subject.create(snapshot) to throw an error if snapshot is invalid.
Catch invalid subjects when appending new subjects to the subject store. Only advance the queue if there are subjects in the store, otherwise we'll get stuck in an infinite loop of appending and advancing.
Export the individual helper functions as named exports from @helpers.
Use valid/invalid instead of success/error in the MIME type tests.