Content pages: remove getInitialProps and the content API (#1635)
This removes `getInitialProps`, and the page content API, in favour of [`getServerSideProps`](https://nextjs.org/docs/basic-features/data-fetching#getserversideprops-server-side-rendering). On each request, page content is fetched from the local content cache and rendered as static HTML (and JSON) for the Team and Publications pages. That HTML  (or JSON) is then served by the app. The content cache (on the server) refreshes from the Contentful API as before.