ZRC: Fix modal default colours (#1786)
* Modal: remove default heading colours
Remove the default heading colours from the Modal component, so that the defaults in ModalHeading are used instead.

* Clean up ModalHeading
Remove the unused Heading component.

* Add color prop to CloseButton
Add a `color` prop to CloseButton. Update stories and add a test.
Pass the Modal title text colour down to the close button.

* Set modal heading size
Set the heading line height to 0 and remove vertical margins, so that the box height is set by its min-height.

* Remove default colours from MovableModal
MovableModal overrides the default brand colours with empty strings, so remove those.

* Fix MovableModal stories
Set the heading background to 'brand' for generic modals. Override the default text colour for subtask popups.

* Add titleColor to SubTaskPopup

* Fix ModalBody padding
Top should be small, medium on the other three sides.

* Add titleColor to ConsensusPopup

* Update React Components changelog