Project app: Subject set picker (#1731)
Add a client-side subject set picker to the workflow selector links, which is shown on click when workflow.grouped is set (ie. when Panoptes expects you to pass a subject set ID with subject queries for that workflow.) Fetch subject sets and subject images with the query that fetches workflows for the project Hero component.