increase the app pods cpu request
Linked to #1794 - increase the pod cpu requests to enable the pods autoscaler a more accurate metric to scale from
e.g. if the request is 10Mi and the app is running at 9Mi (90%) that's already over our 80% target so the the app will scale after sustaining this cpu load for a time period. We don't want this, we want the app to autoscale under more load than this, say 1/4 vCPU (250Mi) for a period of time is our target, e.g. 200Mi load / 250Mi request is 80% cpu use, so we will scale when over the 200Mi value for a time period.