Classifier Dropdown Task (v3) - Simple Dropdown (#1772)
This PR adds a Simple Dropdown (type: dropdown-simple) task to the monorepo Classifier.
The Simple Dropdown functions like a HTML <select> which allows a single value to be selected
The Simple Dropdown is separate from the classic Dropdown (type: dropdown) from PFE
At the moment, the dropdown-simple task type cannot be built in the PFE Project Builder, and has to be set up by a dev.

* Classifier: Simple dropdown task adapter (#1830)

This adds an adapter (and tests) for the simple dropdown task, so that we can test it out in staging. Dropdown tasks are converted to simple dropdown tasks if they contain only one select menu. More complex tasks are not transformed, and should throw an error in the classifier.

Co-authored-by: Jim O'Donnell <jim@zooniverse.org>