Video Data Model (#1889)
* Created SingleVideoSubject

Updated the ‘createLocationCounts’ helper to look for video files. And added the SingleVideoSubject to the SubjectStore.

* Created SingleVideoSubject.spec

Unit test does not pass yet. Still need to create the viewer.
Added singleVideo to the subjectViewers.

* Video Subject

Now detects when to use single video viewer

* Created SingleVideoViewer Component

Just copied SingleImageViewer. Will be going through and updating var names and removing unnecessary code later.
Also added SingleVideoViewer to getViewer.js

* Renamed all SingleVideoViewer file names

and updated index.js and any corresponding import paths.
Fixed error in preload method - somehow ‘fetchImage’ was deleted?! Fat-finger delete?

* Updated video to videos

count.video is an array of videos so renamed to be plural and for clarification.

* More var and file renaming

from image to video

* Removed unit test for files that are WIP

Co-authored-by: Shaun A. Noordin <shaunanoordin@users.noreply.github.com>