Classifier: generic Task connector (#1870)
* Connect Task component to the classifier store
Connect up the Task component so that it gets its annotation and disabled state from the store. Replace useContext and useState with a useStores hook in the Task component. Add an optional store prop for passing stores in during tests.

* Key the task area by classification ID
Unmount and mount the Tasks area for each new classification.