Drawing tools: add actions to handle events during mark creation (#1922)
This extends #1921 to fix #1919 by adding tool.handlePointerMove(event, mark) and tool.handlePointerUp(event, mark) to the tool model. The pointer is captured on mark creation, and pointer events are handled by the drawing surface until the new mark is finished. handlePointerMove and handlePointerUp are called during mark creation and allow tools to implement custom validation of marks. By default, a mark is dragged on pointer move and finished on pointer up, but the transcription line tool overrides these to wait for a second pointer down, which finishes the line. Ellipses might also implement this, to allow for two swipes to create both axes of an ellipse.