ADR: Video Player (#1930)
* Video Player ADR

* Added more details

explaining how react-player is easy to use, easy to style and easy to make responsive.
More descriptive on what files we will allow to be played and where the files will be hosted.
Added explanation why the auto-play feature is not a concern.

* ADR Approved

Added note that all the listed features, ease of styling and ease of customization is compared to the native HTML video player. Updated status to approved.