Video viewer (#1927)
* Update REDAME

update verbage from image to video

* Installed react-player

* Set up SingleVideoViewer

with the react-player component and basic settings. Will be adding to this later.

* Removed all references to ‘draggable’

Not needed for video viewer.

* Removed ‘dragMove’ and ‘setOnDrag’

Not needed for video viewer.

* Removed ‘SVGContext’ and ‘svg’

Not needed for video viewer.

* Removed ‘setOnPan’ and ‘setOnZoom’

Not needed for video viewer.

* Removed ‘enableRotation’

Not needed for video viewer.

* Removed ‘rotation’

Not needed for video viewer.

* Removed ‘move’ and ‘onKeyDown’

Not needed for video viewer.

* Removed ‘SVGPanZoom’

Not needed for video viewer.

* Reorganized some functions

Now they’re in a nice top down, easy to read order!

* Removed ‘fetchImage’ function

And renamed many ‘img’ to ‘vid’

* Removed ‘subjectImageProps’ and ‘subjectImage’

Not needed for video viewer.

* Removed ‘withKeyZoom’

And refactored the export of SingleVideoViewerContainer
withKeyZoom not needed for video viewer.

* Removed ‘enableDrawing’ and ‘enableInteractionLayer’

Not needed for video viewer.

* Video Player appears!

Very basic right now. Stripped a lot of code out and tested along the way to ensure everything is still working. Currently getting a type error for url.  More clean-up and adjusting to come.

* Minor clean up

* Prettier format fixes for VideoViewerContainer

Also added vscode folder to gitignore so all vscode settings and workspace settings do not get committed.

* Final clean up…

for now! Removed console logs. Update ‘vid’ from an object to string. Made a couple Todos as reminders for myself.

* Updated Video Viewer README

* SingleVideoViewer Unit Test

Just testing if component renders properly.

* Prettier formatting

* SingleVideoViewerConnector Unit Test

Updated to simply test render and prop passing.

* Fixed formatting: Replaced tabs with spaces

* Fix more formatting issues

Replaced tab indentation with spacing for all files within the VideoViewer Component.