Add a production test build to CI Tests (#1944)
* Add a production test build to CI Tests
Set up a production build environment. Run a full install (to install the build tools.) Run a test build of all the libraries and the NextJS apps. Run the tests with PANOPTES_ENV=test.
* Content pages: test environment
Specify the environment for logging tests.
* Fix Formik tests
Formik's display name is minimised in production, so find the component instead.
* Mock document.cookie for project UI tests
* Update Readme
Add asset prefixes and Contentful secrets to the Environment Variables section of the Readme.