Remove /classify as the route for a project's default workflow (#1963)
* Workflow selector: remove default link
Link to all workflows by workflow ID in the URL.

* Remove default workflow from projects
project.defaultWorkflow returns either the single active workflow or undefined. If there are multiple workflows, we don't want to choose a workflow for you.

* Convert ProjectHeader to functional components

* Add workflow ID to project navigation
When a project has a single default workflow ID, use that as the URL for the classify page.