Project: Add workflow selection to the Classify page (#1966)
* Add WorkflowSelector to Classify page
Show a workflow selector if there's no workflow ID passed down to the classifier.

* Remove project default workflow
We want people to choose a workflow if a project has more than one workflow, so remove the default selected workflow.

* Check grouped workflows for a selected subject set
Check for a subject set ID if the selected workflow is grouped. If there's no subject set ID, show the workflow selector with the subject set picker active.