Project: Modal workflow selection (#1976)
A continuation of #1966. This rewrites the workflow selector to allow us to have different behaviour on the project home page or on the project classify page. On the classify page, we'd like workflow selection to interrupt you with a modal dialog that forces you to choose a workflow, and optional subject set, before continuing to classify.