Upgrade mobx, mobx-react and mobx-state-tree (#2009)
Update the NextJS apps and the classifier to each use mobx@5.15. 

Bump mobx-react to 6.3 and mobx-state-tree to 3.17 for the NextJS apps.

Rewrite UPPStore tests with mobx.when. 

Upgrade classifier to MST 3.17. 

Tidy up UPP store tests
Move duplicated code into a function. Run tests after the UPP has either loaded or errored, then check that the expected state is set.