Project: Fix all the broken links (#2021)
This extends #2019 to fix broken links to project pages from the dropdown navigation menu (used on small screens or projects with really long names), content boxes (the Stats link) and the project introduction (Learn More.)

The NavLink component is generalised to remove any default link styles and allow any Grommet Anchor prop to be passed in as a prop. DropdownNav, ContentBox and Introduction are updated to replace next/link with the NavLink component. The project Nav component is updated to pass styled Anchor and Text components down to each menu item, now that these aren't the defaults for each link.

Co-authored-by: shaunanoordin <shaunanoordin@users.noreply.github.com>